refactor 광고 수정하기와 만들기의 공통부분 컴포넌트로 분리
